Make GL stuff run with lower GLSL version number
	modified:   ../cvt/gl/GLProgram.cpp
	modified:   ../cvt/gl/GLProgram.h
	modified:   ../cvt/gl/shader/basic.frag
	modified:   ../cvt/gl/shader/text.frag
	modified:   ../cvt/gui/GFXGL.cpp

git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@406 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ cvt/gl/GLProgram.cpp
@@ -80,13 +80,13 @@ namespace cvt {
 	}
 
 
-	void GLProgram::bindAttribLocation( const char *name, size_t location )
+	void GLProgram::bindAttribLocation( const char *name, unsigned int location )
 	{
 	    glBindAttribLocation( program, ( GLuint ) location, name );
 	}
 
 
-	size_t GLProgram::getAttribLocation( const char* name )
+	unsigned int GLProgram::getAttribLocation( const char* name )
 	{
 	    return glGetAttribLocation( program, name );
 	}
